refactor: inline KlaviyoService and UtilService into Pinia stores; delete service files
- Inline all KlaviyoService API calls into klaviyo store (new actions: updateSystemMessageRemote,
  createCommGatewayAuth, updateCommGatewayAuth, deleteCommGatewayAuth, upsertEmailSetting,
  deleteEmailSetting); export maskApiKey, ensureKeyPrefix, generateAuthId as module-level helpers
- Update all call sites (Klaviyo.vue, KlaviyoConnectionDetails.vue, klaviyoEmailEvents.ts) to
  import from @/store/klaviyo instead of @/services/KlaviyoService
- Inline all UtilService API calls into util.ts and netSuite.ts store actions
- Delete src/services/KlaviyoService.ts and src/services/UtilService.ts
